requirements.txt
asgiref==3.8.1
audioread==3.0.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
Django==5.0.6
djangorestframework==3.15.1
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.53.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
matplotlib==3.9.0
moviepy==1.0.3
msgpack==1.0.8
noisereduce==3.0.2
numba==0.59.1
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
pycparser==2.22
pydub==0.25.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
